Entirely rename package to starlings.
