+++ .github/workflows/cd.yml
@@ -28,7 +28,7 @@ jobs:
     - name: Update Rust Cargo.toml version
       run: |
         VERSION=${{ steps.get-version.outputs.version }}
-        sed -i "s/^version = .*/version = \"$VERSION\"/" src/rust/entityframe/Cargo.toml
+        sed -i "s/^version = .*/version = \"$VERSION\"/" src/rust/starlings/Cargo.toml
     
     - name: Upload pyproject.toml
       uses: actions/upload-artifact@v4
@@ -40,7 +40,7 @@ jobs:
       uses: actions/upload-artifact@v4
       with:
         name: cargo-toml
-        path: src/rust/entityframe/Cargo.toml
+        path: src/rust/starlings/Cargo.toml
 
   build-wheels:
     name: Build wheels on ${{ matrix.os }}
@@ -64,7 +64,7 @@ jobs:
       uses: actions/download-artifact@v4
       with:
         name: cargo-toml
-        path: src/rust/entityframe/
+        path: src/rust/starlings/
     
     - name: Install uv and Python
       uses: astral-sh/setup-uv@v6
@@ -110,7 +110,7 @@ jobs:
       uses: actions/download-artifact@v4
       with:
         name: cargo-toml
-        path: src/rust/entityframe/
+        path: src/rust/starlings/
     
     - name: Install uv and Python
       uses: astral-sh/setup-uv@v6
