+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["src/rust/entityframe"]
+members = ["src/rust/starlings"]
 resolver = "2"
