+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "entityframe"
+name = "starlings"
 version = "0.1.0"
 description = "A Python package for comparing entity resolutions from different processes"
 readme = "README.md"
@@ -29,8 +29,8 @@ build-backend = "maturin"
 
 [tool.maturin]
 python-source = "src/python"
-module-name = "entityframe.entityframe"
-manifest-path = "src/rust/entityframe/Cargo.toml"
+module-name = "starlings.starlings"
+manifest-path = "src/rust/starlings/Cargo.toml"
 
 [dependency-groups]
 dev = [
@@ -51,7 +51,7 @@ warn_unused_configs = true
 disallow_untyped_defs = true
 
 [[tool.mypy.overrides]]
-module = "entityframe.entityframe"
+module = "starlings.starlings"
 ignore_missing_imports = true
 
 [tool.pytest.ini_options]
