+++ src/python/starlings/__init__.py
@@ -1,11 +1,11 @@
 """
-EntityFrame - A Python package for comparing entity resolutions from different processes.
+starlings - A Python package for comparing entity resolutions from different processes.
 
 This package provides high-performance entity resolution evaluation using a three-layer
 architecture: string interning, roaring bitmaps, and entity hashing.
 """
 
-from .entityframe import StringInternerCore, EntityCore, CollectionCore
+from .starlings import StringInternerCore, EntityCore, CollectionCore
 from .frame import EntityFrame
 from .entity import Entity
 from .collection import Collection
