+++ src/python/starlings/collection.py
