+++ src/python/starlings/entity.py
