+++ src/rust/starlings/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "entityframe"
+name = "starlings"
 version = "0.1.0"
 edition = "2021"
 
 [lib]
-name = "entityframe"
+name = "starlings"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
