+++ src/rust/starlings/src/hash.rs
