+++ src/rust/starlings/src/lib.rs
@@ -13,7 +13,7 @@ pub use interner::StringInternerCore;
 
 /// A Python module implemented in Rust.
 #[pymodule]
-fn entityframe(m: &Bound<'_, PyModule>) -> PyResult<()> {
+fn starlings(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<StringInternerCore>()?;
     m.add_class::<EntityCore>()?;
     m.add_class::<CollectionCore>()?;
